my_project/requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
bootstrap4==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.42.0
django-bootstrap4==1.1.1
django-crispy-forms==1.9.0
django-filter==2.2.0
django-forms-bootstrap==3.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
oauthlib==3.1.0
packaging==20.4
Pillow==7.1.2
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.7
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
soupsieve==2.0
sqlparse==0.3.1
swagger-ui-py==0.3.0
uritemplate==3.0.1
urllib3==1.25.9
zipp==3.1.0
